+++ cypherast/dialects/puppygraph.py
@@ -9,7 +9,6 @@ from cypherast.dialects.capabilities import DialectCapabilities
 from cypherast.dialects.dialect import register
 from cypherast.dialects.opencypher import OpenCypher, OpenCypherRenderer
 from cypherast.renderer import Renderer
-from cypherast.schema import GraphSchema, modern_graph_schema
 
 
 class PuppyGraphRenderer(OpenCypherRenderer):
@@ -66,19 +65,3 @@ class PuppyGraph(OpenCypher):
         check_undefined_variables=True,
         pattern_predicate_introduces_bindings=False,
     )
-
-    @classmethod
-    def default_schema(cls) -> GraphSchema:
-        """Tutorial person/software schema used when optimize omits ``schema=``."""
-        return modern_graph_schema()
-
-    @classmethod
-    def optimize(
-        cls,
-        tree: a.AstNode,
-        schema: object | None = None,
-        **kwargs: t.Any,
-    ) -> a.AstNode:
-        if schema is None:
-            schema = cls.default_schema()
-        return super().optimize(tree, schema=schema, **kwargs)